Working on forum. Maybe replies should have parents id's, rather than comments having replies lists. Same for (sub)forums.

+++ src/fxControllers/MainWindowController.java
@@ -15,8 +15,11 @@ import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.input.MouseEvent;
 import javafx.stage.Stage;
 
+import javax.swing.event.ListDataListener;
 import java.io.IOException;
 import java.net.URL;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.ResourceBundle;
 
@@ -33,6 +36,14 @@ public class MainWindowController implements Initializable {
     public Button buttonEditUser;
     public Button buttonDeleteUser;
     public Group buttonsGroupTrucks;
+    public Button buttonAddNewTrip;
+    public Button buttonDeleteSelectedTrip;
+    public Group groupButtonsCargo;
+    public TreeView forumTreeView;
+    public TextField textFieldNewForumName;
+    public TextField textFieldCommentTitle;
+    public TextArea textAreaComment;
+    private TreeItem root;
     private User user;
     public Tab usersTab;
     public TableView<TableViewUserParameters> usersTableView;
@@ -42,6 +53,31 @@ public class MainWindowController implements Initializable {
     public TableColumn<TableViewUserParameters, String> roleColumn;
     private ObservableList<TableViewUserParameters> data = FXCollections.observableArrayList();
 
+    @Override
+    public void initialize(URL location, ResourceBundle resources) {
+        nameColumns.setCellValueFactory(new PropertyValueFactory<>("name"));
+        lastNameColumn.setCellValueFactory(new PropertyValueFactory<>("lastName"));
+        roleColumn.setCellValueFactory(new PropertyValueFactory<>("role"));
+        idColumn.setCellValueFactory(new PropertyValueFactory<>("idColumn"));
+
+        Forum root = new Forum();
+        root.setTitle("root cat.");
+        TreeItem rootTreeItem = new TreeItem(root);
+
+        List<Forum> forums = HibernateCRUD.getAllEntity(root);
+        for(Forum forum : forums){
+            rootTreeItem.getChildren().add(new TreeItem(forum));
+        }
+
+        forumTreeView.setRoot(rootTreeItem);
+        rootTreeItem.setExpanded(true);
+    }
+    private <T>TreeItem treeGenerator (TreeItem t){
+        if(t.getValue().getClass().equals(Forum.class)){
+
+        }
+
+    }
     public void setUser(User user) {
         this.user = user;
         if(user.getClass().equals(Manager.class)){
@@ -53,18 +89,27 @@ public class MainWindowController implements Initializable {
         }else{
             usersTab.getTabPane().getTabs().remove(usersTab);
             buttonsGroupTrucks.setVisible(false);
-
+            buttonAddNewTrip.setDisable(true);
+            buttonDeleteSelectedTrip.setDisable(true);
+            groupButtonsCargo.setDisable(true);
         }
         fillTrucksList();
         fillTripsList();
-
-
+        fillCargoList();
     }
 
     private void fillTripsList() {
         Trip trip = new Trip();
         List<Trip> trips = HibernateCRUD.getAllEntity(trip);
-        trips.forEach(c->listViewTrips.getItems().add(c));
+        if(user.getClass().equals(Manager.class)){
+            trips.forEach(c->listViewTrips.getItems().add(c));
+        }else{
+            trips.forEach(c->{
+                if(c.getAssignedDriver()==null || c.getAssignedDriver().equals(user)){
+                    listViewTrips.getItems().add(c);
+                }
+            });
+        }
     }
 
     private void fillUsersLists(){
@@ -84,14 +129,14 @@ public class MainWindowController implements Initializable {
         List<Truck> trucks = HibernateCRUD.getAllEntity(truck);
         trucks.forEach(c->listViewTrucks.getItems().add(c));
     }
-
-    @Override
-    public void initialize(URL location, ResourceBundle resources) {
-        nameColumns.setCellValueFactory(new PropertyValueFactory<>("name"));
-        lastNameColumn.setCellValueFactory(new PropertyValueFactory<>("lastName"));
-        roleColumn.setCellValueFactory(new PropertyValueFactory<>("role"));
-        idColumn.setCellValueFactory(new PropertyValueFactory<>("idColumn"));
+    private void fillCargoList(){
+        listViewCargo.getItems().clear();
+        Cargo cargo = new Cargo();
+        List<Cargo> cargos = HibernateCRUD.getAllEntity(cargo);
+        cargos.forEach(c->listViewCargo.getItems().add(c));
     }
+
+
 
     public void addNewTruck(ActionEvent actionEvent) throws IOException {
         TruckWindowController truckWindowController = new TruckWindowController();
@@ -124,6 +169,8 @@ public class MainWindowController implements Initializable {
     public void addNewTrip(ActionEvent actionEvent) throws IOException {
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxmls/trips-window.fxml"));
         Parent root = loader.load();
+        TripsWindowController tripsWindowController = loader.getController();
+        tripsWindowController.setUser(user);
         Stage stage = new Stage();
         Scene scene = new Scene(root);
         stage.setScene(scene);
@@ -135,6 +182,7 @@ public class MainWindowController implements Initializable {
         Parent root = loader.load();
         TripsWindowController tripsWindowController = loader.getController();
         tripsWindowController.setTrip(listViewTrips.getSelectionModel().getSelectedItem());
+        tripsWindowController.setUser(user);
         tripsWindowController.refreshStopPints();
         tripsWindowController.setUpdateMode();
         Stage stage = new Stage();
@@ -144,9 +192,16 @@ public class MainWindowController implements Initializable {
     }
 
     public void deleteSelectedTrip(ActionEvent actionEvent) {
-        HibernateCRUD.deleteObject(listViewTrips.getSelectionModel().getSelectedItem());
-        listViewTrips.getItems().clear();
-        fillTripsList();
+        if(!(listViewTrips.getSelectionModel().getSelectedItem().isInProcess() || listViewTrips.getSelectionModel().getSelectedItem().isComplete())){
+            HibernateCRUD.deleteObject(listViewTrips.getSelectionModel().getSelectedItem());
+            listViewTrips.getItems().clear();
+            fillTripsList();
+        }else{
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Error");
+            alert.setContentText("Selected trip is already in process or complete.");
+            alert.showAndWait();
+        }
     }
 
     public void refreshTrips(ActionEvent actionEvent) {
@@ -175,7 +230,6 @@ public class MainWindowController implements Initializable {
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxmls/cargo-window.fxml"));
         Parent root = loader.load();
         CargoWindowController cargoWindowController = loader.getController();
-        cargoWindowController.setCargo(listViewCargo.getSelectionModel().getSelectedItem());
         Stage stage = new Stage();
         Scene scene = new Scene(root);
         stage.setScene(scene);
@@ -187,6 +241,7 @@ public class MainWindowController implements Initializable {
         Parent root = loader.load();
         CargoWindowController cargoWindowController = loader.getController();
         cargoWindowController.setCargo(listViewCargo.getSelectionModel().getSelectedItem());
+        System.out.println(listViewCargo.getSelectionModel().getSelectedItem());
         Stage stage = new Stage();
         Scene scene = new Scene(root);
         stage.setScene(scene);
@@ -205,6 +260,45 @@ public class MainWindowController implements Initializable {
                     buttonEditUser.setDisable(true);
                 }
             }
+        }
+    }
+
+    public void mouseClickedOnForum(MouseEvent mouseEvent) {
+    }
+
+    public void addNewSubForum(ActionEvent actionEvent) {
+        Forum forum = new Forum();
+        forum.setTitle(textFieldNewForumName.getText());
+        HibernateCRUD.createObject(forum);
+        Forum selected = ((TreeItem<Forum>) forumTreeView.getSelectionModel().getSelectedItem()).getValue();
+        selected.getSubForums().add(forum);
+        HibernateCRUD.updateObject(selected);
+    }
+
+    public void buttonAddNewComment(ActionEvent actionEvent) {
+        TreeItem<?> treeItem = (TreeItem<?>) forumTreeView.getSelectionModel().getSelectedItem();
+
+        Comment comment = new Comment();
+        comment.setTitle(textFieldCommentTitle.getText());
+        comment.setCommentText(textAreaComment.getText());
+        comment.setAuthor(user);
+        comment.setCreatedDateTime(LocalDateTime.now());
+        HibernateCRUD.createObject(comment);
+
+        if(treeItem.getValue().getClass().equals(Forum.class)){
+            Forum forum = (Forum) treeItem.getValue();
+            forum.getComments().add(comment);
+            HibernateCRUD.updateObject(forum);
+        } else if (treeItem.getValue().getClass().equals(Comment.class)) {
+            Comment selected = (Comment) treeItem.getValue();
+            if(selected.getReplies()!=null){
+                selected.getReplies().add(comment);
+            }else{
+                selected.setReplies(new ArrayList<>());
+                HibernateCRUD.updateObject(selected);
+                selected.getReplies().add(comment);
+            }
+            HibernateCRUD.updateObject(selected);
         }
     }
 }
